Be less eager to raise ENODATA in read
Let one read return no data before raising the ENODATA error.  Some
tools seem to like to check a read at the end of the file, and this
keeps us from throwing errors if they do it once.  It's based on a
global counter, so if different reads get interleaved, it could get
slightly confused.  But there's little harm if this does happen.