Attempt to handle requests for items in the middle of upload
It seems possible that this is a source of trouble in #2.  So we'll
keep a record of all documents being uploaded right now, and check
that if we don't find a document in the cloud or in the buffers.
This could go sideways very quickly if we try to do something with
this document.  But hopefully in practice it will only be setting
attributes or something similarly innocuous.